feat: add HeroContent island for multi-language support
Move the language selection logic from Hero component to a new HeroContent island to improve modularity and reusability. This change enhances maintainability by isolating the language-related functionality into a dedicated component.
